pyproject.toml
[project]
name = "dreamer"
version = "0.1.0"
description = "PyTorch implementation of DreamerV3 - model-based reinforcement learning with learned world models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium[box2d]>=1.2.1",
    "h5py>=3.15.1",
    "hydra-core>=1.3.0",
    "hydra-joblib-launcher>=1.2.0",
    "imageio[ffmpeg]>=2.37.2",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "psutil>=7.1.3",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "swig>=4.3.1.post0",
    "mlflow>=2.10.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
]

[project.scripts]
dreamer-train = "dreamer.main:main"
dreamer-eval = "dreamer.scripts.evaluate:main"
dreamer-viz = "dreamer.scripts.visualize:main"

[tool.uv.sources]
torch = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
